% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Cerebro.R
\name{Cerebro}
\alias{Cerebro}
\title{R6 class in which data sets will be stored for visualization in Cerebro.}
\value{
A new \code{Cerebro} object.
}
\description{
A \code{Cerebro} object is an R6 class that contains several types of
data that can be visualized in Cerebro.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{cerebroApp version that was used to create the object.}

\item{\code{experiment}}{\code{list} that contains meta data about the data set,
including experiment name, species, date of export.}

\item{\code{technical_info}}{\code{list} that contains technical information
about the analysis, including the R session info.}

\item{\code{parameters}}{\code{list} that contains important parameters that
were used during the analysis, e.g. cut-off values for cell filtering.}

\item{\code{groups}}{\code{list} that contains specified grouping variables and
and the group levels (subgroups) that belong to each of them. For each
grouping variable, a corresponding column with the same name must exist
in the meta data.}

\item{\code{cell_cycle}}{\code{vector} that contains the name of columns in the
meta data that contain cell cycle assignments.}

\item{\code{gene_lists}}{\code{list} that contains gene lists, e.g.
mitochondrial and/or ribosomal genes.}

\item{\code{expression}}{\code{matrix}-like object that holds transcript counts.}

\item{\code{meta_data}}{\code{data.frame} that contains cell meta data.}

\item{\code{projections}}{\code{list} that contains projections/dimensional
reductions.}

\item{\code{most_expressed_genes}}{\code{list} that contains a \code{data.frame}
holding the most expressed genes for each grouping variable that was
specified during the call to \code{\link{getMostExpressedGenes}}.}

\item{\code{mean_expression}}{\code{list} that contains a \code{data.frame}
holding the mean expression per gene for each grouping variable.}

\item{\code{marker_genes}}{\code{list} that contains a \code{list} for every
method that was used to calculate marker genes, and a \code{data.frame}
for each grouping variable, e.g. those that were specified during the
call to \code{\link{getMarkerGenes}}.}

\item{\code{enriched_pathways}}{\code{list} that contains a \code{list} for
every method that was used to calculate marker genes, and a
\code{data.frame} for each grouping variable, e.g. those that were
specified during the call to \code{\link{getEnrichedPathways}} or
\code{\link{performGeneSetEnrichmentAnalysis}}.}

\item{\code{trees}}{\code{list} that contains a phylogenetic tree (class
\code{phylo}) for grouping variables.}

\item{\code{trajectories}}{\code{list} that contains a \code{list} for every
method that was used to calculate trajectories, and, depending on the
method, a \code{data.frame} or \code{list} for each specific trajectory,
e.g. those extracted with \code{\link{extractMonocleTrajectory}}.}

\item{\code{extra_material}}{\code{list} that can contain additional material
related to the data set; tables should be stored in \code{data.frame}
format in a named \code{list} called `tables`}

\item{\code{immune_repertoire}}{\code{list} of data.frames (one per sample)
containing scRepertoire columns (CTgene, CTnt, CTaa, CTstrict, etc.).}

\item{\code{bcr_data}}{\code{list} that contains BCR data (kept for backward
compatibility with older .crb files).}

\item{\code{tcr_data}}{\code{list} that contains TCR data (kept for backward
compatibility with older .crb files).}

\item{\code{spatial}}{\code{list} that contains spatial data (coordinates and expression).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cerebro-new}{\code{Cerebro$new()}}
\item \href{#method-Cerebro-setVersion}{\code{Cerebro$setVersion()}}
\item \href{#method-Cerebro-getVersion}{\code{Cerebro$getVersion()}}
\item \href{#method-Cerebro-checkIfGroupExists}{\code{Cerebro$checkIfGroupExists()}}
\item \href{#method-Cerebro-checkIfColumnExistsInMetadata}{\code{Cerebro$checkIfColumnExistsInMetadata()}}
\item \href{#method-Cerebro-addExperiment}{\code{Cerebro$addExperiment()}}
\item \href{#method-Cerebro-getExperiment}{\code{Cerebro$getExperiment()}}
\item \href{#method-Cerebro-addParameters}{\code{Cerebro$addParameters()}}
\item \href{#method-Cerebro-getParameters}{\code{Cerebro$getParameters()}}
\item \href{#method-Cerebro-addTechnicalInfo}{\code{Cerebro$addTechnicalInfo()}}
\item \href{#method-Cerebro-getTechnicalInfo}{\code{Cerebro$getTechnicalInfo()}}
\item \href{#method-Cerebro-addGroup}{\code{Cerebro$addGroup()}}
\item \href{#method-Cerebro-getGroups}{\code{Cerebro$getGroups()}}
\item \href{#method-Cerebro-getGroupLevels}{\code{Cerebro$getGroupLevels()}}
\item \href{#method-Cerebro-setMetaData}{\code{Cerebro$setMetaData()}}
\item \href{#method-Cerebro-getMetaData}{\code{Cerebro$getMetaData()}}
\item \href{#method-Cerebro-addGeneList}{\code{Cerebro$addGeneList()}}
\item \href{#method-Cerebro-getGeneLists}{\code{Cerebro$getGeneLists()}}
\item \href{#method-Cerebro-setExpression}{\code{Cerebro$setExpression()}}
\item \href{#method-Cerebro-getCellNames}{\code{Cerebro$getCellNames()}}
\item \href{#method-Cerebro-getGeneNames}{\code{Cerebro$getGeneNames()}}
\item \href{#method-Cerebro-getMeanExpressionForGenes}{\code{Cerebro$getMeanExpressionForGenes()}}
\item \href{#method-Cerebro-getMeanExpressionForCells}{\code{Cerebro$getMeanExpressionForCells()}}
\item \href{#method-Cerebro-getExpressionMatrix}{\code{Cerebro$getExpressionMatrix()}}
\item \href{#method-Cerebro-setCellCycle}{\code{Cerebro$setCellCycle()}}
\item \href{#method-Cerebro-getCellCycle}{\code{Cerebro$getCellCycle()}}
\item \href{#method-Cerebro-addProjection}{\code{Cerebro$addProjection()}}
\item \href{#method-Cerebro-availableProjections}{\code{Cerebro$availableProjections()}}
\item \href{#method-Cerebro-getProjection}{\code{Cerebro$getProjection()}}
\item \href{#method-Cerebro-addTree}{\code{Cerebro$addTree()}}
\item \href{#method-Cerebro-getTree}{\code{Cerebro$getTree()}}
\item \href{#method-Cerebro-addMostExpressedGenes}{\code{Cerebro$addMostExpressedGenes()}}
\item \href{#method-Cerebro-getGroupsWithMostExpressedGenes}{\code{Cerebro$getGroupsWithMostExpressedGenes()}}
\item \href{#method-Cerebro-getMostExpressedGenes}{\code{Cerebro$getMostExpressedGenes()}}
\item \href{#method-Cerebro-addMeanExpression}{\code{Cerebro$addMeanExpression()}}
\item \href{#method-Cerebro-getGroupsWithMeanExpression}{\code{Cerebro$getGroupsWithMeanExpression()}}
\item \href{#method-Cerebro-getMeanExpression}{\code{Cerebro$getMeanExpression()}}
\item \href{#method-Cerebro-setMarkerGenes}{\code{Cerebro$setMarkerGenes()}}
\item \href{#method-Cerebro-hasMarkerGenes}{\code{Cerebro$hasMarkerGenes()}}
\item \href{#method-Cerebro-getMarkerGenes}{\code{Cerebro$getMarkerGenes()}}
\item \href{#method-Cerebro-getMarkerClusters}{\code{Cerebro$getMarkerClusters()}}
\item \href{#method-Cerebro-addEnrichedPathways}{\code{Cerebro$addEnrichedPathways()}}
\item \href{#method-Cerebro-getMethodsWithEnrichedPathways}{\code{Cerebro$getMethodsWithEnrichedPathways()}}
\item \href{#method-Cerebro-getGroupsWithEnrichedPathways}{\code{Cerebro$getGroupsWithEnrichedPathways()}}
\item \href{#method-Cerebro-getEnrichedPathways}{\code{Cerebro$getEnrichedPathways()}}
\item \href{#method-Cerebro-addTrajectory}{\code{Cerebro$addTrajectory()}}
\item \href{#method-Cerebro-getMethodsWithTrajectories}{\code{Cerebro$getMethodsWithTrajectories()}}
\item \href{#method-Cerebro-getTrajectories}{\code{Cerebro$getTrajectories()}}
\item \href{#method-Cerebro-getTrajectory}{\code{Cerebro$getTrajectory()}}
\item \href{#method-Cerebro-getBCR}{\code{Cerebro$getBCR()}}
\item \href{#method-Cerebro-getTCR}{\code{Cerebro$getTCR()}}
\item \href{#method-Cerebro-addBCRData}{\code{Cerebro$addBCRData()}}
\item \href{#method-Cerebro-addTCRData}{\code{Cerebro$addTCRData()}}
\item \href{#method-Cerebro-getImmuneRepertoire}{\code{Cerebro$getImmuneRepertoire()}}
\item \href{#method-Cerebro-addImmuneRepertoire}{\code{Cerebro$addImmuneRepertoire()}}
\item \href{#method-Cerebro-addSpatialData}{\code{Cerebro$addSpatialData()}}
\item \href{#method-Cerebro-getSpatialData}{\code{Cerebro$getSpatialData()}}
\item \href{#method-Cerebro-availableSpatial}{\code{Cerebro$availableSpatial()}}
\item \href{#method-Cerebro-addExtraMaterial}{\code{Cerebro$addExtraMaterial()}}
\item \href{#method-Cerebro-addExtraTable}{\code{Cerebro$addExtraTable()}}
\item \href{#method-Cerebro-addExtraPlot}{\code{Cerebro$addExtraPlot()}}
\item \href{#method-Cerebro-getExtraMaterial}{\code{Cerebro$getExtraMaterial()}}
\item \href{#method-Cerebro-getExtraMaterialCategories}{\code{Cerebro$getExtraMaterialCategories()}}
\item \href{#method-Cerebro-checkForExtraTables}{\code{Cerebro$checkForExtraTables()}}
\item \href{#method-Cerebro-getNamesOfExtraTables}{\code{Cerebro$getNamesOfExtraTables()}}
\item \href{#method-Cerebro-getExtraTable}{\code{Cerebro$getExtraTable()}}
\item \href{#method-Cerebro-checkForExtraPlots}{\code{Cerebro$checkForExtraPlots()}}
\item \href{#method-Cerebro-getNamesOfExtraPlots}{\code{Cerebro$getNamesOfExtraPlots()}}
\item \href{#method-Cerebro-getExtraPlot}{\code{Cerebro$getExtraPlot()}}
\item \href{#method-Cerebro-print}{\code{Cerebro$print()}}
\item \href{#method-Cerebro-clone}{\code{Cerebro$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Cerebro} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Cerebro} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-setVersion"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-setVersion}{}}}
\subsection{Method \code{setVersion()}}{
Set the version of \code{cerebroApp} that was used to generate this
object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$setVersion(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{Version to set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getVersion"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getVersion}{}}}
\subsection{Method \code{getVersion()}}{
Get the version of \code{cerebroApp} that was used to generate this
object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getVersion()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Version as \code{package_version} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-checkIfGroupExists"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-checkIfGroupExists}{}}}
\subsection{Method \code{checkIfGroupExists()}}{
Safety function that will check if a provided group name is present in
the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$checkIfGroupExists(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name to be tested}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-checkIfColumnExistsInMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-checkIfColumnExistsInMetadata}{}}}
\subsection{Method \code{checkIfColumnExistsInMetadata()}}{
Safety function that will check if a provided group name is present in
the meta data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$checkIfColumnExistsInMetadata(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name to be tested.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addExperiment}{}}}
\subsection{Method \code{addExperiment()}}{
Add information to \code{experiment} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addExperiment(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{organism}.}

\item{\code{content}}{Actual information, e.g. \code{hg}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getExperiment}{}}}
\subsection{Method \code{getExperiment()}}{
Retrieve information from \code{experiment} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getExperiment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{experiment} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addParameters"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addParameters}{}}}
\subsection{Method \code{addParameters()}}{
Add information to \code{parameters} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addParameters(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{number_of_PCs}.}

\item{\code{content}}{Actual information, e.g. \code{30}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getParameters"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getParameters}{}}}
\subsection{Method \code{getParameters()}}{
Retrieve information from \code{parameters} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getParameters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{parameters} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addTechnicalInfo"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addTechnicalInfo}{}}}
\subsection{Method \code{addTechnicalInfo()}}{
Add information to \code{technical_info} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addTechnicalInfo(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{R}.}

\item{\code{content}}{Actual information, e.g. \code{4.0.2}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getTechnicalInfo"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getTechnicalInfo}{}}}
\subsection{Method \code{getTechnicalInfo()}}{
Retrieve information from \code{technical_info} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getTechnicalInfo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{technical_info} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addGroup"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addGroup}{}}}
\subsection{Method \code{addGroup()}}{
Add group to the groups registered in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addGroup(group_name, levels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name.}

\item{\code{levels}}{\code{vector} of group levels (subgroups).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getGroups"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getGroups}{}}}
\subsection{Method \code{getGroups()}}{
Retrieve all names in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getGroups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of registered groups.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getGroupLevels"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getGroupLevels}{}}}
\subsection{Method \code{getGroupLevels()}}{
Retrieve group levels for a group registered in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getGroupLevels(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name for which to retrieve group levels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of group levels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-setMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-setMetaData}{}}}
\subsection{Method \code{setMetaData()}}{
Set meta data for cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$setMetaData(table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{\code{data.frame} that contains meta data for cells. The
number of rows must be equal to the number of rows of projections and
the number of columns in the transcript count matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getMetaData}{}}}
\subsection{Method \code{getMetaData()}}{
Retrieve meta data for cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getMetaData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} containing meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addGeneList"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addGeneList}{}}}
\subsection{Method \code{addGeneList()}}{
Add a gene list to the \code{gene_lists}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addGeneList(name, genes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the gene list.}

\item{\code{genes}}{\code{vector} of genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getGeneLists"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getGeneLists}{}}}
\subsection{Method \code{getGeneLists()}}{
Retrieve gene lists from the \code{gene_lists}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getGeneLists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{gene_lists} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-setExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-setExpression}{}}}
\subsection{Method \code{setExpression()}}{
Set transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$setExpression(counts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{counts}}{\code{matrix}-like object that contains transcript counts
for cells in the data set. Number of columns must be equal to the number
of rows in the \code{meta_data} field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getCellNames"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getCellNames}{}}}
\subsection{Method \code{getCellNames()}}{
Get names of all cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getCellNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing all cell names/barcodes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getGeneNames"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getGeneNames}{}}}
\subsection{Method \code{getGeneNames()}}{
Get names of all genes in transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getGeneNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing all gene names in transcript count matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getMeanExpressionForGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getMeanExpressionForGenes}{}}}
\subsection{Method \code{getMeanExpressionForGenes()}}{
Retrieve mean expression across all cells in the data set for a set of
genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getMeanExpressionForGenes(genes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Names of genes to extract; no default.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing specified gene names and their respective
mean expression across all cells in the data set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getMeanExpressionForCells"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getMeanExpressionForCells}{}}}
\subsection{Method \code{getMeanExpressionForCells()}}{
Retrieve (mean) expression for a single gene or a set of genes for a
given set of cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getMeanExpressionForCells(cells = NULL, genes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{Names/barcodes of cells to extract; defaults to \code{NULL},
which will return all cells.}

\item{\code{genes}}{Names of genes to extract; defaults to \code{NULL}, which
will return all genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} containing (mean) expression across all specified genes in
each specified cell.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getExpressionMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getExpressionMatrix}{}}}
\subsection{Method \code{getExpressionMatrix()}}{
Retrieve transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getExpressionMatrix(cells = NULL, genes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{Names/barcodes of cells to extract; defaults to \code{NULL},
which will return all cells.}

\item{\code{genes}}{Names of genes to extract; defaults to \code{NULL}, which
will return all genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dense transcript count matrix for specified cells and genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-setCellCycle"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-setCellCycle}{}}}
\subsection{Method \code{setCellCycle()}}{
Add columns containing cell cycle assignments to the \code{cell_cycle}
field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$setCellCycle(cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{\code{vector} of columns names containing cell cycle
assignments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getCellCycle"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getCellCycle}{}}}
\subsection{Method \code{getCellCycle()}}{
Retrieve column names containing cell cycle assignments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getCellCycle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of column names in meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addProjection"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addProjection}{}}}
\subsection{Method \code{addProjection()}}{
Add projections (dimensional reductions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addProjection(name, projection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the projection.}

\item{\code{projection}}{\code{data.frame} containing positions of cells in
projection.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-availableProjections"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-availableProjections}{}}}
\subsection{Method \code{availableProjections()}}{
Get list of available projections (dimensional reductions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$availableProjections()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of projections / dimensional reductions that are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getProjection"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getProjection}{}}}
\subsection{Method \code{getProjection()}}{
Retrieve data for a specific projection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getProjection(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of projection.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the positions of cells in the projection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addTree"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addTree}{}}}
\subsection{Method \code{addTree()}}{
Add phylogenetic tree to \code{trees} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addTree(group_name, tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name that this tree belongs to.}

\item{\code{tree}}{Phylogenetic tree as \code{phylo} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getTree"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getTree}{}}}
\subsection{Method \code{getTree()}}{
Retrieve phylogenetic tree for a specific group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getTree(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name for which to retrieve phylogenetic tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Phylogenetic tree as \code{phylo} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addMostExpressedGenes}{}}}
\subsection{Method \code{addMostExpressedGenes()}}{
Add table of most expressed genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addMostExpressedGenes(group_name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable that the most expressed genes
belong to. Must be registered in the \code{groups} field.}

\item{\code{table}}{\code{data.frame} that contains the most expressed genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getGroupsWithMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getGroupsWithMostExpressedGenes}{}}}
\subsection{Method \code{getGroupsWithMostExpressedGenes()}}{
Retrieve names of grouping variables for which most expressed genes are
available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getGroupsWithMostExpressedGenes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of grouping variables for which most expressed genes are
available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getMostExpressedGenes}{}}}
\subsection{Method \code{getMostExpressedGenes()}}{
Retrieve table of most expressed genes for a specific grouping variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getMostExpressedGenes(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable for which to retrieve most
expressed genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the most expressed genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addMeanExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addMeanExpression}{}}}
\subsection{Method \code{addMeanExpression()}}{
Add table of mean expression per gene.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addMeanExpression(group_name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable that the mean expression
belongs to. Must be registered in the \code{groups} field.}

\item{\code{table}}{\code{data.frame} that contains the mean expression per gene.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getGroupsWithMeanExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getGroupsWithMeanExpression}{}}}
\subsection{Method \code{getGroupsWithMeanExpression()}}{
Retrieve names of grouping variables for which mean expression data is
available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getGroupsWithMeanExpression()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of grouping variables for which mean expression is
available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getMeanExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getMeanExpression}{}}}
\subsection{Method \code{getMeanExpression()}}{
Retrieve table of mean expression for a specific grouping variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getMeanExpression(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable for which to retrieve mean
expression.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the mean expression per gene.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-setMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-setMarkerGenes}{}}}
\subsection{Method \code{setMarkerGenes()}}{
Set marker genes table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$setMarkerGenes(table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{\code{data.frame} that contains the marker genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-hasMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-hasMarkerGenes}{}}}
\subsection{Method \code{hasMarkerGenes()}}{
Check if marker genes are available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$hasMarkerGenes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} indicating whether marker genes are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getMarkerGenes}{}}}
\subsection{Method \code{getMarkerGenes()}}{
Retrieve marker genes table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getMarkerGenes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} containing the marker genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getMarkerClusters"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getMarkerClusters}{}}}
\subsection{Method \code{getMarkerClusters()}}{
Get unique cluster/group values from marker genes table.
Looks for common column names like "group", "cluster", "cell_type", etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getMarkerClusters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of unique cluster/group values, or \code{NULL} if not found.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addEnrichedPathways}{}}}
\subsection{Method \code{addEnrichedPathways()}}{
Add table of enriched pathways.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addEnrichedPathways(method, group_name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to calculate enriched
pathways.}

\item{\code{group_name}}{Name of grouping variable that the enriched pathways
belong to. Must be registered in the \code{groups} field.}

\item{\code{table}}{\code{data.frame} that contains the enriched pathways.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getMethodsWithEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getMethodsWithEnrichedPathways}{}}}
\subsection{Method \code{getMethodsWithEnrichedPathways()}}{
Retrieve names of methods for which enriched pathways are available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getMethodsWithEnrichedPathways()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of methods for which enriched pathways are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getGroupsWithEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getGroupsWithEnrichedPathways}{}}}
\subsection{Method \code{getGroupsWithEnrichedPathways()}}{
Retrieve names of grouping variables for which enriched pathways are
available for a specific method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getGroupsWithEnrichedPathways(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method for which to retrieve grouping variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of grouping variables for which enriched pathways are
available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getEnrichedPathways}{}}}
\subsection{Method \code{getEnrichedPathways()}}{
Retrieve table of enriched pathways for a specific method and grouping
variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getEnrichedPathways(method, group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method for which to retrieve enriched pathways.}

\item{\code{group_name}}{Name of grouping variable for which to retrieve enriched
pathways.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the enriched pathways.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addTrajectory"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addTrajectory}{}}}
\subsection{Method \code{addTrajectory()}}{
Add trajectory to \code{trajectories} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addTrajectory(method, trajectory_name, trajectory)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to calculate trajectory.}

\item{\code{trajectory_name}}{Name of trajectory.}

\item{\code{trajectory}}{Trajectory data as \code{data.frame} or \code{list}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getMethodsWithTrajectories"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getMethodsWithTrajectories}{}}}
\subsection{Method \code{getMethodsWithTrajectories()}}{
Retrieve names of methods for which trajectories are available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getMethodsWithTrajectories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of methods for which trajectories are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getTrajectories"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getTrajectories}{}}}
\subsection{Method \code{getTrajectories()}}{
Retrieve names of trajectories for a specific method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getTrajectories(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method for which to retrieve trajectories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of trajectories for the specified method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getTrajectory"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getTrajectory}{}}}
\subsection{Method \code{getTrajectory()}}{
Retrieve trajectory data for a specific method and trajectory name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getTrajectory(method, trajectory_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method for which to retrieve trajectory.}

\item{\code{trajectory_name}}{Name of trajectory to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Trajectory data as \code{data.frame} or \code{list}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getBCR"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getBCR}{}}}
\subsection{Method \code{getBCR()}}{
Retrieve BCR data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getBCR()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BCR data stored in the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getTCR"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getTCR}{}}}
\subsection{Method \code{getTCR()}}{
Retrieve TCR data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getTCR()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TCR data stored in the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addBCRData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addBCRData}{}}}
\subsection{Method \code{addBCRData()}}{
Add BCR data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addBCRData(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{list} that contains BCR data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addTCRData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addTCRData}{}}}
\subsection{Method \code{addTCRData()}}{
Add TCR data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addTCRData(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{list} that contains TCR data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getImmuneRepertoire"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getImmuneRepertoire}{}}}
\subsection{Method \code{getImmuneRepertoire()}}{
Get immune repertoire data. Returns the unified \code{immune_repertoire}
field if available; otherwise falls back to merging legacy
\code{bcr_data} and \code{tcr_data}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getImmuneRepertoire()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named list of data.frames (one per sample), or empty list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addImmuneRepertoire"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addImmuneRepertoire}{}}}
\subsection{Method \code{addImmuneRepertoire()}}{
Set immune repertoire data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addImmuneRepertoire(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Named list of data.frames (one per sample) containing
scRepertoire columns.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addSpatialData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addSpatialData}{}}}
\subsection{Method \code{addSpatialData()}}{
Add spatial data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addSpatialData(name, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the spatial data entry (e.g. image name).}

\item{\code{data}}{\code{list} containing 'coordinates' (data.frame) and 'expression' (sparse matrix).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getSpatialData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getSpatialData}{}}}
\subsection{Method \code{getSpatialData()}}{
Retrieve spatial data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getSpatialData(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the spatial data entry.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} containing 'coordinates' and 'expression'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-availableSpatial"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-availableSpatial}{}}}
\subsection{Method \code{availableSpatial()}}{
Get list of available spatial data entries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$availableSpatial()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of spatial data entries that are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addExtraMaterial"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addExtraMaterial}{}}}
\subsection{Method \code{addExtraMaterial()}}{
Add content to extra material field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addExtraMaterial(category, name, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{Name of category. At the moment, only \code{tables} and
\code{plots} are valid categories. Tables must be in \code{data.frame}
format and plots must be created with \code{ggplot2}.}

\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{content}}{Data that should be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addExtraTable"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addExtraTable}{}}}
\subsection{Method \code{addExtraTable()}}{
Add table to `extra_material` slot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addExtraTable(name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{table}}{Table that should be added, must be \code{data.frame}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-addExtraPlot"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-addExtraPlot}{}}}
\subsection{Method \code{addExtraPlot()}}{
Add plot to `extra_material` slot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$addExtraPlot(name, plot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{plot}}{Plot that should be added, must be created with
\code{ggplot2} (class: \code{ggplot}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getExtraMaterial"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getExtraMaterial}{}}}
\subsection{Method \code{getExtraMaterial()}}{
Retrieve extra material from \code{extra_material} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getExtraMaterial()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{extra_material} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getExtraMaterialCategories"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getExtraMaterialCategories}{}}}
\subsection{Method \code{getExtraMaterialCategories()}}{
Get names of categories for which extra material is available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getExtraMaterialCategories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} with names of available categories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-checkForExtraTables"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-checkForExtraTables}{}}}
\subsection{Method \code{checkForExtraTables()}}{
Check whether there are tables in the extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$checkForExtraTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} indicating whether there are tables in the extra
materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getNamesOfExtraTables"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getNamesOfExtraTables}{}}}
\subsection{Method \code{getNamesOfExtraTables()}}{
Get names of tables in extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getNamesOfExtraTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing names of tables in extra materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getExtraTable"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getExtraTable}{}}}
\subsection{Method \code{getExtraTable()}}{
Get table from extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getExtraTable(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Requested table in \code{data.frame} format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-checkForExtraPlots"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-checkForExtraPlots}{}}}
\subsection{Method \code{checkForExtraPlots()}}{
Check whether there are plots in the extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$checkForExtraPlots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} indicating whether there are plots in the extra
materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getNamesOfExtraPlots"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getNamesOfExtraPlots}{}}}
\subsection{Method \code{getNamesOfExtraPlots()}}{
Get names of plots in extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getNamesOfExtraPlots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing names of plots in extra materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-getExtraPlot"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-getExtraPlot}{}}}
\subsection{Method \code{getExtraPlot()}}{
Get plot from extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$getExtraPlot(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Requested plot made with \code{ggplot2}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-print"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-print}{}}}
\subsection{Method \code{print()}}{
Show overview of object and the data it contains.
Extract expression matrix (helper)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
