% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Cerebro_v1.3.R
\name{Cerebro_v1.3}
\alias{Cerebro_v1.3}
\title{R6 class in which data sets will be stored for visualization in Cerebro.}
\value{
A new \code{Cerebro_v1.3} object.
}
\description{
A \code{Cerebro_v1.3} object is an R6 class that contains several types of
data that can be visualized in Cerebro.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{version}}{cerebroApp version that was used to create the object.}

\item{\code{experiment}}{\code{list} that contains meta data about the data set,
including experiment name, species, date of export.}

\item{\code{technical_info}}{\code{list} that contains technical information
about the analysis, including the R session info.}

\item{\code{parameters}}{\code{list} that contains important parameters that
were used during the analysis, e.g. cut-off values for cell filtering.}

\item{\code{groups}}{\code{list} that contains specified grouping variables and
and the group levels (subgroups) that belong to each of them. For each
grouping variable, a corresponding column with the same name must exist
in the meta data.}

\item{\code{cell_cycle}}{\code{vector} that contains the name of columns in the
meta data that contain cell cycle assignments.}

\item{\code{gene_lists}}{\code{list} that contains gene lists, e.g.
mitochondrial and/or ribosomal genes.}

\item{\code{expression}}{\code{matrix}-like object that holds transcript counts.}

\item{\code{expression_backend}}{\code{list} describing how/where the expression
matrix is stored. For step 7.1 every newly exported object tags itself
\code{list(type = "embedded", location = NULL)}; future step 7.2 will
introduce \code{type = "h5"} / \code{"bpcells"} with an external
\code{location}. Older \code{.crb} files (serialised before this field
existed) load with \code{expression_backend = NULL}; \code{getExpressionBackend()}
treats that as \code{"embedded"} for backward compatibility.}

\item{\code{meta_data}}{\code{data.frame} that contains cell meta data.}

\item{\code{projections}}{\code{list} that contains projections/dimensional
reductions.}

\item{\code{most_expressed_genes}}{\code{list} that contains a \code{data.frame}
holding the most expressed genes for each grouping variable that was
specified during the call to \code{\link{getMostExpressedGenes}}.}

\item{\code{mean_expression}}{\code{list} that contains a \code{data.frame}
holding the mean expression per gene for each grouping variable.}

\item{\code{marker_genes}}{\code{list} that contains a \code{list} for every
method that was used to calculate marker genes, and a \code{data.frame}
for each grouping variable, e.g. those that were specified during the
call to \code{\link{getMarkerGenes}}.}

\item{\code{enriched_pathways}}{\code{list} that contains a \code{list} for
every method that was used to calculate marker genes, and a
\code{data.frame} for each grouping variable, e.g. those that were
specified during the call to \code{\link{getEnrichedPathways}} or
\code{\link{performGeneSetEnrichmentAnalysis}}.}

\item{\code{trees}}{\code{list} that contains a phylogenetic tree (class
\code{phylo}) for grouping variables.}

\item{\code{trajectories}}{\code{list} that contains a \code{list} for every
method that was used to calculate trajectories, and, depending on the
method, a \code{data.frame} or \code{list} for each specific trajectory,
e.g. those extracted with \code{\link{extractMonocleTrajectory}}.}

\item{\code{extra_material}}{\code{list} that can contain additional material
related to the data set; tables should be stored in \code{data.frame}
format in a named \code{list} called `tables`}

\item{\code{immune_repertoire}}{\code{list} of data.frames (one per sample)
containing scRepertoire columns (CTgene, CTnt, CTaa, CTstrict, etc.).}

\item{\code{bcr_data}}{\code{list} that contains BCR data (kept for backward
compatibility with older .crb files).}

\item{\code{tcr_data}}{\code{list} that contains TCR data (kept for backward
compatibility with older .crb files).}

\item{\code{spatial}}{\code{list} that contains spatial data (coordinates and expression).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Cerebro_v1.3-new}{\code{Cerebro_v1.3$new()}}
\item \href{#method-Cerebro_v1.3-setVersion}{\code{Cerebro_v1.3$setVersion()}}
\item \href{#method-Cerebro_v1.3-getVersion}{\code{Cerebro_v1.3$getVersion()}}
\item \href{#method-Cerebro_v1.3-checkIfGroupExists}{\code{Cerebro_v1.3$checkIfGroupExists()}}
\item \href{#method-Cerebro_v1.3-checkIfColumnExistsInMetadata}{\code{Cerebro_v1.3$checkIfColumnExistsInMetadata()}}
\item \href{#method-Cerebro_v1.3-addExperiment}{\code{Cerebro_v1.3$addExperiment()}}
\item \href{#method-Cerebro_v1.3-getExperiment}{\code{Cerebro_v1.3$getExperiment()}}
\item \href{#method-Cerebro_v1.3-addParameters}{\code{Cerebro_v1.3$addParameters()}}
\item \href{#method-Cerebro_v1.3-getParameters}{\code{Cerebro_v1.3$getParameters()}}
\item \href{#method-Cerebro_v1.3-addTechnicalInfo}{\code{Cerebro_v1.3$addTechnicalInfo()}}
\item \href{#method-Cerebro_v1.3-getTechnicalInfo}{\code{Cerebro_v1.3$getTechnicalInfo()}}
\item \href{#method-Cerebro_v1.3-addGroup}{\code{Cerebro_v1.3$addGroup()}}
\item \href{#method-Cerebro_v1.3-getGroups}{\code{Cerebro_v1.3$getGroups()}}
\item \href{#method-Cerebro_v1.3-getGroupLevels}{\code{Cerebro_v1.3$getGroupLevels()}}
\item \href{#method-Cerebro_v1.3-setMetaData}{\code{Cerebro_v1.3$setMetaData()}}
\item \href{#method-Cerebro_v1.3-getMetaData}{\code{Cerebro_v1.3$getMetaData()}}
\item \href{#method-Cerebro_v1.3-addGeneList}{\code{Cerebro_v1.3$addGeneList()}}
\item \href{#method-Cerebro_v1.3-getGeneLists}{\code{Cerebro_v1.3$getGeneLists()}}
\item \href{#method-Cerebro_v1.3-setExpression}{\code{Cerebro_v1.3$setExpression()}}
\item \href{#method-Cerebro_v1.3-setExpressionBackend}{\code{Cerebro_v1.3$setExpressionBackend()}}
\item \href{#method-Cerebro_v1.3-getExpressionBackend}{\code{Cerebro_v1.3$getExpressionBackend()}}
\item \href{#method-Cerebro_v1.3-getCellNames}{\code{Cerebro_v1.3$getCellNames()}}
\item \href{#method-Cerebro_v1.3-getGeneNames}{\code{Cerebro_v1.3$getGeneNames()}}
\item \href{#method-Cerebro_v1.3-getMeanExpressionForGenes}{\code{Cerebro_v1.3$getMeanExpressionForGenes()}}
\item \href{#method-Cerebro_v1.3-getMeanExpressionForCells}{\code{Cerebro_v1.3$getMeanExpressionForCells()}}
\item \href{#method-Cerebro_v1.3-getExpressionMatrix}{\code{Cerebro_v1.3$getExpressionMatrix()}}
\item \href{#method-Cerebro_v1.3-getExpressionRow}{\code{Cerebro_v1.3$getExpressionRow()}}
\item \href{#method-Cerebro_v1.3-getExpressionBlock}{\code{Cerebro_v1.3$getExpressionBlock()}}
\item \href{#method-Cerebro_v1.3-setCellCycle}{\code{Cerebro_v1.3$setCellCycle()}}
\item \href{#method-Cerebro_v1.3-getCellCycle}{\code{Cerebro_v1.3$getCellCycle()}}
\item \href{#method-Cerebro_v1.3-addProjection}{\code{Cerebro_v1.3$addProjection()}}
\item \href{#method-Cerebro_v1.3-availableProjections}{\code{Cerebro_v1.3$availableProjections()}}
\item \href{#method-Cerebro_v1.3-getProjection}{\code{Cerebro_v1.3$getProjection()}}
\item \href{#method-Cerebro_v1.3-addTree}{\code{Cerebro_v1.3$addTree()}}
\item \href{#method-Cerebro_v1.3-getTree}{\code{Cerebro_v1.3$getTree()}}
\item \href{#method-Cerebro_v1.3-addMostExpressedGenes}{\code{Cerebro_v1.3$addMostExpressedGenes()}}
\item \href{#method-Cerebro_v1.3-getGroupsWithMostExpressedGenes}{\code{Cerebro_v1.3$getGroupsWithMostExpressedGenes()}}
\item \href{#method-Cerebro_v1.3-getMostExpressedGenes}{\code{Cerebro_v1.3$getMostExpressedGenes()}}
\item \href{#method-Cerebro_v1.3-addMeanExpression}{\code{Cerebro_v1.3$addMeanExpression()}}
\item \href{#method-Cerebro_v1.3-getGroupsWithMeanExpression}{\code{Cerebro_v1.3$getGroupsWithMeanExpression()}}
\item \href{#method-Cerebro_v1.3-getMeanExpression}{\code{Cerebro_v1.3$getMeanExpression()}}
\item \href{#method-Cerebro_v1.3-addMarkerGenes}{\code{Cerebro_v1.3$addMarkerGenes()}}
\item \href{#method-Cerebro_v1.3-getMethodsForMarkerGenes}{\code{Cerebro_v1.3$getMethodsForMarkerGenes()}}
\item \href{#method-Cerebro_v1.3-getGroupsWithMarkerGenes}{\code{Cerebro_v1.3$getGroupsWithMarkerGenes()}}
\item \href{#method-Cerebro_v1.3-getMarkerGenes}{\code{Cerebro_v1.3$getMarkerGenes()}}
\item \href{#method-Cerebro_v1.3-addEnrichedPathways}{\code{Cerebro_v1.3$addEnrichedPathways()}}
\item \href{#method-Cerebro_v1.3-getMethodsWithEnrichedPathways}{\code{Cerebro_v1.3$getMethodsWithEnrichedPathways()}}
\item \href{#method-Cerebro_v1.3-getGroupsWithEnrichedPathways}{\code{Cerebro_v1.3$getGroupsWithEnrichedPathways()}}
\item \href{#method-Cerebro_v1.3-getEnrichedPathways}{\code{Cerebro_v1.3$getEnrichedPathways()}}
\item \href{#method-Cerebro_v1.3-addTrajectory}{\code{Cerebro_v1.3$addTrajectory()}}
\item \href{#method-Cerebro_v1.3-getMethodsForTrajectories}{\code{Cerebro_v1.3$getMethodsForTrajectories()}}
\item \href{#method-Cerebro_v1.3-getNamesOfTrajectories}{\code{Cerebro_v1.3$getNamesOfTrajectories()}}
\item \href{#method-Cerebro_v1.3-getTrajectory}{\code{Cerebro_v1.3$getTrajectory()}}
\item \href{#method-Cerebro_v1.3-getBCR}{\code{Cerebro_v1.3$getBCR()}}
\item \href{#method-Cerebro_v1.3-getTCR}{\code{Cerebro_v1.3$getTCR()}}
\item \href{#method-Cerebro_v1.3-addBCRData}{\code{Cerebro_v1.3$addBCRData()}}
\item \href{#method-Cerebro_v1.3-addTCRData}{\code{Cerebro_v1.3$addTCRData()}}
\item \href{#method-Cerebro_v1.3-getImmuneRepertoire}{\code{Cerebro_v1.3$getImmuneRepertoire()}}
\item \href{#method-Cerebro_v1.3-addImmuneRepertoire}{\code{Cerebro_v1.3$addImmuneRepertoire()}}
\item \href{#method-Cerebro_v1.3-addSpatialData}{\code{Cerebro_v1.3$addSpatialData()}}
\item \href{#method-Cerebro_v1.3-getSpatialData}{\code{Cerebro_v1.3$getSpatialData()}}
\item \href{#method-Cerebro_v1.3-availableSpatial}{\code{Cerebro_v1.3$availableSpatial()}}
\item \href{#method-Cerebro_v1.3-addExtraMaterial}{\code{Cerebro_v1.3$addExtraMaterial()}}
\item \href{#method-Cerebro_v1.3-addExtraTable}{\code{Cerebro_v1.3$addExtraTable()}}
\item \href{#method-Cerebro_v1.3-addExtraPlot}{\code{Cerebro_v1.3$addExtraPlot()}}
\item \href{#method-Cerebro_v1.3-getExtraMaterial}{\code{Cerebro_v1.3$getExtraMaterial()}}
\item \href{#method-Cerebro_v1.3-getExtraMaterialCategories}{\code{Cerebro_v1.3$getExtraMaterialCategories()}}
\item \href{#method-Cerebro_v1.3-checkForExtraTables}{\code{Cerebro_v1.3$checkForExtraTables()}}
\item \href{#method-Cerebro_v1.3-getNamesOfExtraTables}{\code{Cerebro_v1.3$getNamesOfExtraTables()}}
\item \href{#method-Cerebro_v1.3-getExtraTable}{\code{Cerebro_v1.3$getExtraTable()}}
\item \href{#method-Cerebro_v1.3-checkForExtraPlots}{\code{Cerebro_v1.3$checkForExtraPlots()}}
\item \href{#method-Cerebro_v1.3-getNamesOfExtraPlots}{\code{Cerebro_v1.3$getNamesOfExtraPlots()}}
\item \href{#method-Cerebro_v1.3-getExtraPlot}{\code{Cerebro_v1.3$getExtraPlot()}}
\item \href{#method-Cerebro_v1.3-print}{\code{Cerebro_v1.3$print()}}
\item \href{#method-Cerebro_v1.3-clone}{\code{Cerebro_v1.3$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-new"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Cerebro_v1.3} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Cerebro_v1.3} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-setVersion"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-setVersion}{}}}
\subsection{Method \code{setVersion()}}{
Set the version of \code{cerebroApp} that was used to generate this
object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$setVersion(version)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{version}}{Version to set.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getVersion"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getVersion}{}}}
\subsection{Method \code{getVersion()}}{
Get the version of \code{cerebroApp} that was used to generate this
object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getVersion()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Version as \code{package_version} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-checkIfGroupExists"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-checkIfGroupExists}{}}}
\subsection{Method \code{checkIfGroupExists()}}{
Safety function that will check if a provided group name is present in
the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$checkIfGroupExists(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name to be tested}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-checkIfColumnExistsInMetadata"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-checkIfColumnExistsInMetadata}{}}}
\subsection{Method \code{checkIfColumnExistsInMetadata()}}{
Safety function that will check if a provided group name is present in
the meta data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$checkIfColumnExistsInMetadata(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name to be tested.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addExperiment}{}}}
\subsection{Method \code{addExperiment()}}{
Add information to \code{experiment} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addExperiment(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{organism}.}

\item{\code{content}}{Actual information, e.g. \code{hg}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getExperiment"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getExperiment}{}}}
\subsection{Method \code{getExperiment()}}{
Retrieve information from \code{experiment} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExperiment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{experiment} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addParameters"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addParameters}{}}}
\subsection{Method \code{addParameters()}}{
Add information to \code{parameters} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addParameters(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{number_of_PCs}.}

\item{\code{content}}{Actual information, e.g. \code{30}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getParameters"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getParameters}{}}}
\subsection{Method \code{getParameters()}}{
Retrieve information from \code{parameters} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getParameters()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{parameters} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addTechnicalInfo"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addTechnicalInfo}{}}}
\subsection{Method \code{addTechnicalInfo()}}{
Add information to \code{technical_info} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addTechnicalInfo(field, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field}}{Name of the information, e.g. \code{R}.}

\item{\code{content}}{Actual information, e.g. \code{4.0.2}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getTechnicalInfo"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getTechnicalInfo}{}}}
\subsection{Method \code{getTechnicalInfo()}}{
Retrieve information from \code{technical_info} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getTechnicalInfo()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{technical_info} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addGroup"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addGroup}{}}}
\subsection{Method \code{addGroup()}}{
Add group to the groups registered in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addGroup(group_name, levels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name.}

\item{\code{levels}}{\code{vector} of group levels (subgroups).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getGroups"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getGroups}{}}}
\subsection{Method \code{getGroups()}}{
Retrieve all names in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of registered groups.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getGroupLevels"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getGroupLevels}{}}}
\subsection{Method \code{getGroupLevels()}}{
Retrieve group levels for a group registered in the \code{groups} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroupLevels(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name for which to retrieve group levels.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of group levels.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-setMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-setMetaData}{}}}
\subsection{Method \code{setMetaData()}}{
Set meta data for cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$setMetaData(table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{\code{data.frame} that contains meta data for cells. The
number of rows must be equal to the number of rows of projections and
the number of columns in the transcript count matrix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getMetaData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getMetaData}{}}}
\subsection{Method \code{getMetaData()}}{
Retrieve meta data for cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMetaData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} containing meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addGeneList"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addGeneList}{}}}
\subsection{Method \code{addGeneList()}}{
Add a gene list to the \code{gene_lists}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addGeneList(name, genes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the gene list.}

\item{\code{genes}}{\code{vector} of genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getGeneLists"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getGeneLists}{}}}
\subsection{Method \code{getGeneLists()}}{
Retrieve gene lists from the \code{gene_lists}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGeneLists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{gene_lists} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-setExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-setExpression}{}}}
\subsection{Method \code{setExpression()}}{
Set transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$setExpression(counts, backend = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{counts}}{\code{matrix}-like object that contains transcript counts
for cells in the data set. Number of columns must be equal to the number
of rows in the \code{meta_data} field.}

\item{\code{backend}}{Optional backend tag. If left \code{NULL} the object is
tagged \code{"embedded"} (the matrix lives inside the \code{.crb}
itself). Callers exporting with step-7.2 external-storage modes should
pass \code{setExpressionBackend()} directly instead of relying on this
default.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-setExpressionBackend"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-setExpressionBackend}{}}}
\subsection{Method \code{setExpressionBackend()}}{
Tag the object with information about how / where its expression matrix
is stored. In step 7.1 every newly exported \code{.crb} is tagged
\code{"embedded"} with a NULL location, meaning the matrix is carried
inside the serialised \code{.crb}. Later steps (7.2 exporter, 7.3
runtime attach) will produce objects tagged \code{"h5"} or
\code{"bpcells"} with an external \code{location}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$setExpressionBackend(type = "embedded", location = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Storage backend label. One of \code{"embedded"},
\code{"h5"}, \code{"bpcells"}. Step 7.1 only recognises
\code{"embedded"} at runtime; the other two are accepted here (so step
7.2 can set them) but will still need step-7.3 runtime attach to be
useful.}

\item{\code{location}}{Optional character path (absolute or relative to the
generated app \code{data/} directory) where the external matrix lives.
\code{NULL} when \code{type == "embedded"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getExpressionBackend"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getExpressionBackend}{}}}
\subsection{Method \code{getExpressionBackend()}}{
Read the expression backend tag. Returns a \code{list(type, location)}.
For \code{.crb} files generated before the \code{expression_backend}
field existed the stored slot is \code{NULL}; this method graciously
falls back to \code{list(type = "embedded", location = NULL)} so that
downstream code does not need to special-case legacy objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExpressionBackend()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getCellNames"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getCellNames}{}}}
\subsection{Method \code{getCellNames()}}{
Get names of all cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getCellNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing all cell names/barcodes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getGeneNames"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getGeneNames}{}}}
\subsection{Method \code{getGeneNames()}}{
Get names of all genes in transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGeneNames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing all gene names in transcript count matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getMeanExpressionForGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getMeanExpressionForGenes}{}}}
\subsection{Method \code{getMeanExpressionForGenes()}}{
Retrieve mean expression across all cells in the data set for a set of
genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMeanExpressionForGenes(genes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Names of genes to extract; no default.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing specified gene names and their respective
mean expression across all cells in the data set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getMeanExpressionForCells"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getMeanExpressionForCells}{}}}
\subsection{Method \code{getMeanExpressionForCells()}}{
Retrieve (mean) expression for a single gene or a set of genes for a
given set of cells.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMeanExpressionForCells(cells = NULL, genes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{Names/barcodes of cells to extract; defaults to \code{NULL},
which will return all cells.}

\item{\code{genes}}{Names of genes to extract; defaults to \code{NULL}, which
will return all genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} containing (mean) expression across all specified genes in
each specified cell.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getExpressionMatrix"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getExpressionMatrix}{}}}
\subsection{Method \code{getExpressionMatrix()}}{
Retrieve transcript count matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExpressionMatrix(cells = NULL, genes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cells}}{Names/barcodes of cells to extract; defaults to \code{NULL},
which will return all cells.}

\item{\code{genes}}{Names of genes to extract; defaults to \code{NULL}, which
will return all genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Dense transcript count matrix for specified cells and genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getExpressionRow"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getExpressionRow}{}}}
\subsection{Method \code{getExpressionRow()}}{
Retrieve a single row of the expression matrix as a named numeric vector
WITHOUT going through the dense helper. Prefer this over
\code{getExpressionMatrix(genes = gene)} on large or sparse backends
where materialising a 1 x N dense matrix first is wasteful.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExpressionRow(gene, cells = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gene}}{Name of a single gene. Must exist in the matrix.}

\item{\code{cells}}{Names/barcodes of cells to extract; \code{NULL} returns all cells.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Named \code{numeric} vector, one entry per requested cell.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getExpressionBlock"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getExpressionBlock}{}}}
\subsection{Method \code{getExpressionBlock()}}{
Retrieve a genes x cells sub-matrix in the backend's NATIVE form
(sparse / lazy). Callers that need a dense base matrix must apply
\code{as.matrix()} themselves. Use this to keep sparse-aware downstream
operations (\code{Matrix::rowMeans}, \code{Matrix::colMeans}, etc.)
fast instead of densifying just to aggregate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExpressionBlock(genes, cells = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{genes}}{Non-empty character vector of gene names.}

\item{\code{cells}}{Names/barcodes of cells to extract; \code{NULL} returns all cells.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A sub-matrix of the same concrete class as \code{self$expression}:
\code{dgCMatrix} stays \code{dgCMatrix}, \code{RleMatrix} yields
\code{DelayedMatrix}, \code{IterableMatrix} stays \code{IterableMatrix}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-setCellCycle"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-setCellCycle}{}}}
\subsection{Method \code{setCellCycle()}}{
Add columns containing cell cycle assignments to the \code{cell_cycle}
field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$setCellCycle(cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cols}}{\code{vector} of columns names containing cell cycle
assignments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getCellCycle"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getCellCycle}{}}}
\subsection{Method \code{getCellCycle()}}{
Retrieve column names containing cell cycle assignments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getCellCycle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of column names in meta data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addProjection"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addProjection}{}}}
\subsection{Method \code{addProjection()}}{
Add projections (dimensional reductions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addProjection(name, projection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the projection.}

\item{\code{projection}}{\code{data.frame} containing positions of cells in
projection.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-availableProjections"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-availableProjections}{}}}
\subsection{Method \code{availableProjections()}}{
Get list of available projections (dimensional reductions).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$availableProjections()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of projections / dimensional reductions that are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getProjection"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getProjection}{}}}
\subsection{Method \code{getProjection()}}{
Retrieve data for a specific projection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getProjection(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of projection.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the positions of cells in the projection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addTree"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addTree}{}}}
\subsection{Method \code{addTree()}}{
Add phylogenetic tree to \code{trees} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addTree(group_name, tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name that this tree belongs to.}

\item{\code{tree}}{Phylogenetic tree as \code{phylo} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getTree"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getTree}{}}}
\subsection{Method \code{getTree()}}{
Retrieve phylogenetic tree for a specific group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getTree(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Group name for which to retrieve phylogenetic tree.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Phylogenetic tree as \code{phylo} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addMostExpressedGenes}{}}}
\subsection{Method \code{addMostExpressedGenes()}}{
Add table of most expressed genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addMostExpressedGenes(group_name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable that the most expressed genes
belong to. Must be registered in the \code{groups} field.}

\item{\code{table}}{\code{data.frame} that contains the most expressed genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getGroupsWithMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getGroupsWithMostExpressedGenes}{}}}
\subsection{Method \code{getGroupsWithMostExpressedGenes()}}{
Retrieve names of grouping variables for which most expressed genes are
available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroupsWithMostExpressedGenes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of grouping variables for which most expressed genes are
available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getMostExpressedGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getMostExpressedGenes}{}}}
\subsection{Method \code{getMostExpressedGenes()}}{
Retrieve table of most expressed genes for a specific grouping variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMostExpressedGenes(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable for which to retrieve most
expressed genes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the most expressed genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addMeanExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addMeanExpression}{}}}
\subsection{Method \code{addMeanExpression()}}{
Add table of mean expression per gene.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addMeanExpression(group_name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable that the mean expression
belongs to. Must be registered in the \code{groups} field.}

\item{\code{table}}{\code{data.frame} that contains the mean expression per gene.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getGroupsWithMeanExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getGroupsWithMeanExpression}{}}}
\subsection{Method \code{getGroupsWithMeanExpression()}}{
Retrieve names of grouping variables for which mean expression data is
available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroupsWithMeanExpression()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of grouping variables for which mean expression is
available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getMeanExpression"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getMeanExpression}{}}}
\subsection{Method \code{getMeanExpression()}}{
Retrieve table of mean expression for a specific grouping variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMeanExpression(group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_name}}{Name of grouping variable for which to retrieve mean
expression.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the mean expression per gene.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addMarkerGenes}{}}}
\subsection{Method \code{addMarkerGenes()}}{
Add table of marker genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addMarkerGenes(method, name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to generate the marker genes.}

\item{\code{name}}{Name of table. This name will be used to select the table in
Cerebro. It is recommended to use the grouping variable, e.g.
\code{sample}.}

\item{\code{table}}{\code{data.frame} that contains the marker genes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getMethodsForMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getMethodsForMarkerGenes}{}}}
\subsection{Method \code{getMethodsForMarkerGenes()}}{
Retrieve names of methods that were used to generate marker genes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMethodsForMarkerGenes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of names of methods that were used to generate marker
genes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getGroupsWithMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getGroupsWithMarkerGenes}{}}}
\subsection{Method \code{getGroupsWithMarkerGenes()}}{
Retrieve grouping variables for which marker genes were generated using
a specified method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroupsWithMarkerGenes(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of grouping variables for which marker genes were
calculated using the specified method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getMarkerGenes"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getMarkerGenes}{}}}
\subsection{Method \code{getMarkerGenes()}}{
Retrieve table of marker genes for specific method and grouping variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMarkerGenes(method, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method.}

\item{\code{name}}{Name of table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} that contains marker genes for the specified
combination of method and grouping variable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addEnrichedPathways}{}}}
\subsection{Method \code{addEnrichedPathways()}}{
Add table of enriched pathways.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addEnrichedPathways(method, group_name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to calculate enriched
pathways.}

\item{\code{group_name}}{Name of grouping variable that the enriched pathways
belong to. Must be registered in the \code{groups} field.}

\item{\code{table}}{\code{data.frame} that contains the enriched pathways.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getMethodsWithEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getMethodsWithEnrichedPathways}{}}}
\subsection{Method \code{getMethodsWithEnrichedPathways()}}{
Retrieve names of methods for which enriched pathways are available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMethodsWithEnrichedPathways()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of methods for which enriched pathways are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getGroupsWithEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getGroupsWithEnrichedPathways}{}}}
\subsection{Method \code{getGroupsWithEnrichedPathways()}}{
Retrieve names of grouping variables for which enriched pathways are
available for a specific method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getGroupsWithEnrichedPathways(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method for which to retrieve grouping variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of grouping variables for which enriched pathways are
available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getEnrichedPathways"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getEnrichedPathways}{}}}
\subsection{Method \code{getEnrichedPathways()}}{
Retrieve table of enriched pathways for a specific method and grouping
variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getEnrichedPathways(method, group_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method for which to retrieve enriched pathways.}

\item{\code{group_name}}{Name of grouping variable for which to retrieve enriched
pathways.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the enriched pathways.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addTrajectory"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addTrajectory}{}}}
\subsection{Method \code{addTrajectory()}}{
Add trajectory to \code{trajectories} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addTrajectory(method, trajectory_name, trajectory)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method that was used to calculate trajectory.}

\item{\code{trajectory_name}}{Name of trajectory.}

\item{\code{trajectory}}{Trajectory data as \code{data.frame} or \code{list}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getMethodsForTrajectories"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getMethodsForTrajectories}{}}}
\subsection{Method \code{getMethodsForTrajectories()}}{
Retrieve names of methods for which trajectories are available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getMethodsForTrajectories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of methods for which trajectories are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getNamesOfTrajectories"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getNamesOfTrajectories}{}}}
\subsection{Method \code{getNamesOfTrajectories()}}{
Retrieve names of trajectories for a specific method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getNamesOfTrajectories(method)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method for which to retrieve trajectories.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{vector} of trajectories for the specified method.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getTrajectory"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getTrajectory}{}}}
\subsection{Method \code{getTrajectory()}}{
Retrieve trajectory data for a specific method and trajectory name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getTrajectory(method, trajectory_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{Name of method for which to retrieve trajectory.}

\item{\code{trajectory_name}}{Name of trajectory to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Trajectory data as \code{data.frame} or \code{list}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getBCR"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getBCR}{}}}
\subsection{Method \code{getBCR()}}{
Retrieve BCR data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getBCR()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
BCR data stored in the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getTCR"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getTCR}{}}}
\subsection{Method \code{getTCR()}}{
Retrieve TCR data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getTCR()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TCR data stored in the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addBCRData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addBCRData}{}}}
\subsection{Method \code{addBCRData()}}{
Add BCR data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addBCRData(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{list} that contains BCR data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addTCRData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addTCRData}{}}}
\subsection{Method \code{addTCRData()}}{
Add TCR data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addTCRData(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{\code{list} that contains TCR data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getImmuneRepertoire"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getImmuneRepertoire}{}}}
\subsection{Method \code{getImmuneRepertoire()}}{
Get immune repertoire data. Returns the unified \code{immune_repertoire}
field if available; otherwise falls back to merging legacy
\code{bcr_data} and \code{tcr_data}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getImmuneRepertoire()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named list of data.frames (one per sample), or empty list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addImmuneRepertoire"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addImmuneRepertoire}{}}}
\subsection{Method \code{addImmuneRepertoire()}}{
Set immune repertoire data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addImmuneRepertoire(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Named list of data.frames (one per sample) containing
scRepertoire columns.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addSpatialData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addSpatialData}{}}}
\subsection{Method \code{addSpatialData()}}{
Add spatial data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addSpatialData(name, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the spatial data entry (e.g. image name).}

\item{\code{data}}{\code{list} containing 'coordinates' (data.frame) and 'expression' (sparse matrix).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getSpatialData"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getSpatialData}{}}}
\subsection{Method \code{getSpatialData()}}{
Retrieve spatial data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getSpatialData(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the spatial data entry.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} containing 'coordinates' and 'expression'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-availableSpatial"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-availableSpatial}{}}}
\subsection{Method \code{availableSpatial()}}{
Get list of available spatial data entries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$availableSpatial()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} of spatial data entries that are available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addExtraMaterial"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addExtraMaterial}{}}}
\subsection{Method \code{addExtraMaterial()}}{
Add content to extra material field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addExtraMaterial(category, name, content)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{Name of category. At the moment, only \code{tables} and
\code{plots} are valid categories. Tables must be in \code{data.frame}
format and plots must be created with \code{ggplot2}.}

\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{content}}{Data that should be added.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addExtraTable"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addExtraTable}{}}}
\subsection{Method \code{addExtraTable()}}{
Add table to `extra_material` slot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addExtraTable(name, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{table}}{Table that should be added, must be \code{data.frame}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-addExtraPlot"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-addExtraPlot}{}}}
\subsection{Method \code{addExtraPlot()}}{
Add plot to `extra_material` slot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$addExtraPlot(name, plot)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of material, will be used to select it in Cerebro.}

\item{\code{plot}}{Plot that should be added, must be created with
\code{ggplot2} (class: \code{ggplot}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getExtraMaterial"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getExtraMaterial}{}}}
\subsection{Method \code{getExtraMaterial()}}{
Retrieve extra material from \code{extra_material} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExtraMaterial()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of all entries in the \code{extra_material} field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getExtraMaterialCategories"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getExtraMaterialCategories}{}}}
\subsection{Method \code{getExtraMaterialCategories()}}{
Get names of categories for which extra material is available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExtraMaterialCategories()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} with names of available categories.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-checkForExtraTables"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-checkForExtraTables}{}}}
\subsection{Method \code{checkForExtraTables()}}{
Check whether there are tables in the extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$checkForExtraTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} indicating whether there are tables in the extra
materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getNamesOfExtraTables"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getNamesOfExtraTables}{}}}
\subsection{Method \code{getNamesOfExtraTables()}}{
Get names of tables in extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getNamesOfExtraTables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing names of tables in extra materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getExtraTable"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getExtraTable}{}}}
\subsection{Method \code{getExtraTable()}}{
Get table from extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExtraTable(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Requested table in \code{data.frame} format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-checkForExtraPlots"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-checkForExtraPlots}{}}}
\subsection{Method \code{checkForExtraPlots()}}{
Check whether there are plots in the extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$checkForExtraPlots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical} indicating whether there are plots in the extra
materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getNamesOfExtraPlots"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getNamesOfExtraPlots}{}}}
\subsection{Method \code{getNamesOfExtraPlots()}}{
Get names of plots in extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getNamesOfExtraPlots()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{vector} containing names of plots in extra materials.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-getExtraPlot"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-getExtraPlot}{}}}
\subsection{Method \code{getExtraPlot()}}{
Get plot from extra materials.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$getExtraPlot(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Requested plot made with \code{ggplot2}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-print"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-print}{}}}
\subsection{Method \code{print()}}{
Show overview of object and the data it contains.
Print overview of available marker gene results for \code{self$print()}
function.
Print overview of available enriched pathway results for
\code{self$print()} function.
Print overview of available trajectories for \code{self$print()} function.
Print overview of extra material for \code{self$print()} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cerebro_v1.3-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cerebro_v1.3-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cerebro_v1.3$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
