% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_typing.R
\name{hla_allele_locus}
\alias{hla_allele_locus}
\title{Locus of a canonical allele (e.g. "HLA-A")}
\usage{
hla_allele_locus(allele)
}
\description{
Locus of a canonical allele (e.g. "HLA-A")
}
\keyword{internal}
