% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_typing.R
\name{hla_allele_resolution}
\alias{hla_allele_resolution}
\title{Count the field resolution of a canonical allele (e.g. 2-field)}
\usage{
hla_allele_resolution(allele)
}
\description{
Count the field resolution of a canonical allele (e.g. 2-field)
}
\keyword{internal}
