% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_typing.R
\name{hla_locus_class}
\alias{hla_locus_class}
\title{MHC class of a locus: "Class I" / "Class II" / "Other"}
\usage{
hla_locus_class(locus)
}
\description{
MHC class of a locus: "Class I" / "Class II" / "Other"
}
\keyword{internal}
